Warn about ignored before-notifications
This sent me down quite the rabbit hole today, because this patched
version silently ignores :before-notifications. This patch at least adds
a warning to all chef logs that :before notifications are ignored here,
which makes it easier to troubleshoot issues with this.

This change will also mitigate #66 by telling the user. It's no fix, but
it's less confusing this way.